WEB_SkillSwap_REST/requirements.txt
aiosqlite
alembic
fastapi
pydantic
python-dotenv
SQLAlchemy
uvicorn
asyncpg
psycopg2
